Lots of things
Moved data graphing to its own component.

Styled some CSS.

Changed graphical data updates to be more 'react-like'.
